+++ src/useOnchainKit.test.tsx
@@ -2,45 +2,19 @@
  * @jest-environment jsdom
  */
 import React from 'react';
-import { render, screen, waitFor } from '@testing-library/react';
-import { OnchainKitProvider, useOnchainKit } from './useOnchainKit'; // Adjust the import path as necessary
-import { EASSchemaUid } from './identity/types';
-
 import '@testing-library/jest-dom';
+import { render } from '@testing-library/react';
+import { useOnchainKit } from './useOnchainKit';
 
 const TestComponent = () => {
   const { identity } = useOnchainKit();
   return <div>{identity.eas.schemaId}</div>;
 };
 
-describe('OnchainKitProvider and useOnchainKit', () => {
-  it('provides the context value correctly', async () => {
-    const schemaId: EASSchemaUid = `0x${'1'.repeat(64)}`;
-
-    render(
-      <OnchainKitProvider identity={{ easConfig: { schemaId } }}>
-        <TestComponent />
-      </OnchainKitProvider>,
-    );
-
-    await waitFor(() => {
-      expect(screen.getByText(schemaId)).toBeInTheDocument();
-    });
-  });
-
-  it('throws an error if schemaId does not meet the required length', () => {
-    expect(() =>
-      render(
-        <OnchainKitProvider identity={{ easConfig: { schemaId: '0x123' } }}>
-          <TestComponent />
-        </OnchainKitProvider>,
-      ),
-    ).toThrow('EAS schemaId must be 64 characters prefixed with "0x"');
-  });
-
+describe('useOnchainKit', () => {
   it('throws an error when useOnchainKit is used outside of OnchainKitProvider', () => {
     expect(() => render(<TestComponent />)).toThrow(
-      'useIdentity must be used within an OnchainKitProvider',
+      'useOnchainKit must be used within an OnchainKitProvider',
     );
   });
 });
